Refactor MD5 utility to decouple GUI dependencies
